Update S benchmark suite version
This commit adds the updated version of the S benchmark suite

+++ benchmark-suite/fairness/fairness.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 # Copyright (C) 2013 Paolo Valente <paolo.valente@unimore.it>
 #                    Arianna Avanzini <avanzini.arianna@gmail.com>
+
+../utilities/check_dependencies.sh awk dd fio iostat
+if [[ $? -ne 0 ]]; then
+	exit
+fi
+
 . ../config_params.sh
 . ../utilities/lib_utils.sh
 CALC_AVG_AND_CO=`cd ../utilities; pwd`/calc_avg_and_co.sh
@@ -17,13 +23,14 @@ switches to bfq and launches 10 iterations of 2 sequential readers of 2 \n\
 different files of 100MB each; the first reader has weight 1000, the second\n\
 500.\n\
 \n\
-Default parameters values are bfq, 4, 2, $NUM_BLOCKS, and 100 for each reader\n"
+Default parameter values are bfq, 4, 2, $NUM_BLOCKS, and 100 for every reader\n"
 
 SCHED=${1-bfq}
 NUM_FILES=${2-4}
 ITERATIONS=${3-2}
 NUM_BLOCKS=${4-$NUM_BLOCKS}
 R_TYPE=${5-seq}
+BFQ_NEW_VERSION=Y
 
 if [ "$1" == "-h" ]; then
 	printf "$usage_msg"
@@ -32,9 +39,16 @@ fi
 
 # set proper group
 if [ "${SCHED}" == "bfq" ] ; then
+    if [ "${BFQ_NEW_VERSION}" == "Y" ]; then
+	GROUP="blkio"
+	PREFIX="bfq."
+    else
 	GROUP="bfqio"
+	PREFIX=""
+    fi
 elif [ "${SCHED}" == "cfq" ] ; then
 	GROUP="blkio"
+	PREFIX=""
 fi
 
 mkdir -p /cgroup
@@ -52,19 +66,25 @@ for ((i = 0 ; $i < $NUM_FILES ; i++)) ; do
 	fi
 done
 
+# create files to read
+create_files_rw_type $NUM_FILES $R_TYPE
+
 # initialize weight array
+echo -n "Weights:"
 args=("$@")
 max_w=${WEIGHT[0]}
 for ((i = 0 ; $i < $NUM_FILES ; i++)) ; do
-	if [ "${args[$(($i+6))]}" != "" ] ; then
-		WEIGHT[$i]=${args[$(($i+6))]}
+	if [ "${args[$(($i+5))]}" != "" ] ; then
+		WEIGHT[$i]=${args[$(($i+5))]}
 	else
 		WEIGHT[$i]=100
 	fi
 	if [[ ${WEIGHT[$i]} -gt $max_w ]] ; then
 		max_w=${WEIGHT[$i]}
 	fi
+	echo -n " ${WEIGHT[$i]}"
 done
+echo
 
 # create result dir tree and cd to its root
 rm -rf results-${SCHED}
@@ -126,7 +146,7 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 	idx=0
 	echo $FILES
 	for f in $FILES ; do
-		echo ${WEIGHT[$idx]} > /cgroup/test$idx/$GROUP.weight
+		echo ${WEIGHT[$idx]} > /cgroup/test$idx/$GROUP.${PREFIX}weight
 		create_reader_and_assign_to_group $R_TYPE $i $idx $f
 		idx=$(($idx+1))
 	done
@@ -137,7 +157,7 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 	# start logging aggregated throughput
 	iostat -tmd /dev/$HD 5 | tee iter-$i/iostat.out &
 
-	if [[ "$TIMEOUT" != "0" ]]; then
+	if [[ "$TIMEOUT" != "0" && "$TIMEOUT" != "" ]]; then
 		bash -c "sleep $TIMEOUT && \
 			 echo Timeout: killing readers ;\
 			 killall -q -s USR1 dd ; sleep 1 ;\
@@ -152,7 +172,8 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 		wait %$j
 	done
 
-	if [[ "$KILLPROC" != "" && "$(ps $KILLPROC | tail -n +2)" != "" ]]; then
+	if [[ "$KILLPROC" != "" && "$(ps $KILLPROC | tail -n +2)" != "" ]];
+	then
 		kill -9 $KILLPROC > /dev/null 2>&1
 	fi
 	KILLPROC=
@@ -160,6 +181,13 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 done
 
 set_tracing 0
+
+# destroy cgroups and unmount controller
+for ((i = 0 ; $i < $NUM_FILES ; i++)) ; do
+	rmdir /cgroup/test$i
+done
+umount /cgroup
+rm -rf /cgroup
 
 for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 	cd iter-$i
@@ -172,7 +200,8 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 
 	echo reader time stats in iteration $i | tee -a ../output
 	if [[ "$R_TYPE" == "seq" ]]; then
-		cat singles/* | grep "copied\|copiati" | awk '{ print $6 }' > time
+		cat singles/* | grep "copied\|copiati" | awk '{ print $6 }' \
+		    > time
 	else
 		rm -f time; touch time
 		for s in $(ls -1 singles/*); do
@@ -186,7 +215,8 @@ for ((i = 0 ; $i < $ITERATIONS ; i++)) ; do
 
 	echo reader bandwith stats in iteration $i | tee -a ../output
 	if [[ "$R_TYPE" == "seq" ]]; then
-		cat singles/* | grep "copied\|copiati" | awk '{ print $8 }' > band
+		cat singles/* | grep "copied\|copiati" | awk '{ print $8 }' \
+		    > band
 	else
 		rm -f band; touch band
 		for s in $(ls -1 singles/*); do
